Cast r,g and b to int in operator<< for XYZRGB type so that they are human readable.
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@4275 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ common/include/pcl/impl/point_types.hpp
@@ -431,7 +431,7 @@ namespace pcl
   };
   inline std::ostream& operator << (std::ostream& os, const PointXYZRGB& p)
   {
-    os << "(" << p.x << "," << p.y << "," << p.z << " - " << p.r << "," << p.g << "," << p.b << ")";
+    os << "(" << p.x << "," << p.y << "," << p.z << " - " << (int) p.r << "," << (int) p.g << "," << (int) p.b << ")";
     return (os);
   }
